v0.9.1: fixed incorrect int arithmetic when artifacts transfer levels

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/Artifact.java
@@ -130,7 +130,7 @@ public class Artifact extends KindofMisc {
 
 	//transfers upgrades from another artifact, transfer level will equal the displayed level
 	public void transferUpgrade(int transferLvl) {
-		upgrade(Math.round((float)(transferLvl*levelCap)/10));
+		upgrade(Math.round((transferLvl*levelCap)/10f));
 	}
 
 	@Override
